requirements.txt
streamlit==1.29.0
streamlit-authenticator==0.2.3
extra-streamlit-components==0.1.60
firebase-admin==6.2.0
python-dotenv==1.0.0
openai==1.6.0
requests==2.31.0
pandas==2.0.3
pillow==10.0.0
matplotlib==3.7.3
pyyaml==6.0.1
bcrypt==4.0.1